Move bootstrap phase processing to bootstrap manager
