+++ ggdeploymentd/src/bootstrap_manager.c
@@ -5,16 +5,19 @@
 #include "bootstrap_manager.h"
 #include "deployment_model.h"
 #include "deployment_queue.h"
+#include <fcntl.h>
 #include <ggl/buffer.h>
 #include <ggl/bump_alloc.h>
 #include <ggl/core_bus/gg_config.h>
 #include <ggl/error.h>
+#include <ggl/file.h>
 #include <ggl/log.h>
 #include <ggl/map.h>
 #include <ggl/object.h>
 #include <ggl/vector.h>
 #include <stdbool.h>
 #include <stdint.h>
+#include <stdlib.h>
 
 bool bootstrap_required(GglMap recipe, GglBuffer component_name) {
     GglObject *lifecycle_config;
@@ -78,6 +81,9 @@ GglError save_deployment_state(
               deploymentType: local/IoT Jobs
               deploymentDoc:
     */
+
+    GGL_LOGD("Encountered component requiring bootstrap. Saving deployment "
+             "state to config.");
 
     GglObject deployment_doc = GGL_OBJ_MAP(GGL_MAP(
         { GGL_STR("deployment_id"), GGL_OBJ_BUF(deployment->deployment_id) },
@@ -177,6 +183,8 @@ GglError save_deployment_state(
 GglError retrieve_in_progress_deployment(
     GglDeployment *deployment, GglKVVec *deployed_components
 ) {
+    GGL_LOGD("Searching config for any in progress deployment.");
+
     GglBuffer config_mem = GGL_BUF((uint8_t[2500]) { 0 });
     GglBumpAlloc balloc = ggl_bump_alloc_init(config_mem);
     GglObject deployment_config;
@@ -349,6 +357,8 @@ GglError retrieve_in_progress_deployment(
 }
 
 GglError delete_saved_deployment_from_config(void) {
+    GGL_LOGD("Deleting previously saved deployment from config.");
+
     GglError ret = ggl_gg_config_delete(GGL_BUF_LIST(
         GGL_STR("services"),
         GGL_STR("DeploymentService"),
@@ -363,4 +373,163 @@ GglError delete_saved_deployment_from_config(void) {
     }
 
     return GGL_ERR_OK;
+}
+
+GglError process_bootstrap_phase(
+    GglMap components,
+    GglBuffer root_path,
+    GglBufVec *bootstrap_comp_name_buf_vec
+) {
+    GGL_MAP_FOREACH(component, components) {
+        GglBuffer component_name = component->key;
+
+        static uint8_t bootstrap_service_file_path_buf[PATH_MAX];
+        GglByteVec bootstrap_service_file_path_vec
+            = GGL_BYTE_VEC(bootstrap_service_file_path_buf);
+        GglError ret
+            = ggl_byte_vec_append(&bootstrap_service_file_path_vec, root_path);
+        ggl_byte_vec_append(&bootstrap_service_file_path_vec, GGL_STR("/"));
+        ggl_byte_vec_append(&bootstrap_service_file_path_vec, GGL_STR("ggl."));
+        ggl_byte_vec_chain_append(
+            &ret, &bootstrap_service_file_path_vec, component_name
+        );
+        ggl_byte_vec_chain_append(
+            &ret,
+            &bootstrap_service_file_path_vec,
+            GGL_STR(".bootstrap.service")
+        );
+        if (ret == GGL_ERR_OK) {
+            // check if the current component name has relevant bootstrap
+            // service file created
+            int fd = -1;
+            ret = ggl_file_open(
+                bootstrap_service_file_path_vec.buf, O_RDONLY, 0, &fd
+            );
+            if (ret != GGL_ERR_OK) {
+                GGL_LOGD(
+                    "Component %.*s does not have the relevant bootstrap "
+                    "service file",
+                    (int) component_name.len,
+                    component_name.data
+                );
+            } else { // relevant bootstrap service file exists
+
+                // add relevant component name into the vector
+                ret = ggl_buf_vec_push(
+                    bootstrap_comp_name_buf_vec, component_name
+                );
+                if (ret != GGL_ERR_OK) {
+                    GGL_LOGE("Failed to add the bootstrap component name "
+                             "into vector");
+                    return ret;
+                }
+
+                // initiate link command for 'bootstrap'
+                static uint8_t link_command_buf[PATH_MAX];
+                GglByteVec link_command_vec = GGL_BYTE_VEC(link_command_buf);
+                ret = ggl_byte_vec_append(
+                    &link_command_vec, GGL_STR("systemctl link ")
+                );
+                ggl_byte_vec_chain_append(
+                    &ret, &link_command_vec, bootstrap_service_file_path_vec.buf
+                );
+                ggl_byte_vec_chain_push(&ret, &link_command_vec, '\0');
+                if (ret != GGL_ERR_OK) {
+                    GGL_LOGE(
+                        "Failed to create systemctl link command for:%.*s",
+                        (int) bootstrap_service_file_path_vec.buf.len,
+                        bootstrap_service_file_path_vec.buf.data
+                    );
+                    return ret;
+                }
+
+                GGL_LOGD(
+                    "Command to execute: %.*s",
+                    (int) link_command_vec.buf.len,
+                    link_command_vec.buf.data
+                );
+
+                // NOLINTBEGIN(concurrency-mt-unsafe)
+                int system_ret = system((char *) link_command_vec.buf.data);
+                if (WIFEXITED(system_ret)) {
+                    if (WEXITSTATUS(system_ret) != 0) {
+                        GGL_LOGE(
+                            "systemctl link failed for:%.*s",
+                            (int) bootstrap_service_file_path_vec.buf.len,
+                            bootstrap_service_file_path_vec.buf.data
+                        );
+                        return ret;
+                    }
+                    GGL_LOGI(
+                        "systemctl link exited for %.*s with child status "
+                        "%d\n",
+                        (int) bootstrap_service_file_path_vec.buf.len,
+                        bootstrap_service_file_path_vec.buf.data,
+                        WEXITSTATUS(system_ret)
+                    );
+                } else {
+                    GGL_LOGE(
+                        "systemctl link did not exit normally for %.*s",
+                        (int) bootstrap_service_file_path_vec.buf.len,
+                        bootstrap_service_file_path_vec.buf.data
+                    );
+                    return ret;
+                }
+
+                // initiate start command for 'bootstrap'
+                static uint8_t start_command_buf[PATH_MAX];
+                GglByteVec start_command_vec = GGL_BYTE_VEC(start_command_buf);
+                ret = ggl_byte_vec_append(
+                    &start_command_vec, GGL_STR("systemctl start ")
+                );
+                ggl_byte_vec_chain_append(
+                    &ret, &start_command_vec, GGL_STR("ggl.")
+                );
+                ggl_byte_vec_chain_append(
+                    &ret, &start_command_vec, component_name
+                );
+                ggl_byte_vec_chain_append(
+                    &ret, &start_command_vec, GGL_STR(".bootstrap.service\0")
+                );
+
+                GGL_LOGD(
+                    "Command to execute: %.*s",
+                    (int) start_command_vec.buf.len,
+                    start_command_vec.buf.data
+                );
+                if (ret != GGL_ERR_OK) {
+                    GGL_LOGE(
+                        "Failed to create systemctl start command for %.*s",
+                        (int) bootstrap_service_file_path_vec.buf.len,
+                        bootstrap_service_file_path_vec.buf.data
+                    );
+                    return ret;
+                }
+
+                system_ret = system((char *) start_command_vec.buf.data);
+                // NOLINTEND(concurrency-mt-unsafe)
+                if (WIFEXITED(system_ret)) {
+                    if (WEXITSTATUS(system_ret) != 0) {
+                        GGL_LOGE(
+                            "systemctl start failed for%.*s",
+                            (int) bootstrap_service_file_path_vec.buf.len,
+                            bootstrap_service_file_path_vec.buf.data
+                        );
+                        return ret;
+                    }
+                    GGL_LOGI(
+                        "systemctl start exited with child status %d\n",
+                        WEXITSTATUS(system_ret)
+                    );
+                } else {
+                    GGL_LOGE(
+                        "systemctl start did not exit normally for %.*s",
+                        (int) bootstrap_service_file_path_vec.buf.len,
+                        bootstrap_service_file_path_vec.buf.data
+                    );
+                    return ret;
+                }
+            }
+        }
+    }
 }
